Crate namada_governance

Governance library code

Re-exports

pub use storage::proposal::InitProposalData;

pub use storage::proposal::ProposalType;

pub use storage::proposal::VoteProposalData;

pub use storage::vote::ProposalVote;

pub use storage::init_proposal;

pub use storage::is_proposal_accepted;

pub use storage::vote_proposal;

Modules

cli

governance CLI structures

event

Governance transaction events.

parameters

governance parameters

pgf

governance public good fundings PGF library code

storage

governance storage Governance storage

utils

Governance utility functions/structs

vp

Governance VP

Structs

Store

Governance storage Keys/Read/Write implementation

Constants

ADDRESS

The governance internal address

Traits

Read

Abstract governance storage read interface

Write

Abstract governance storage write interface

Functions

finalize_block

Apply governance updates for a block. On a new epoch, this will look for proposals to tally completed proposals and execute accepted proposals.

Type Aliases

Result

Result of a storage API call.