namada_governance::pgf::storage
Module keys

Pgf storage keys

Functions

fundings_handle

LazyMap handler for the pgf fundings substorage

fundings_key_prefix

Obtain a storage key for pgf fundings.

get_fundings_key_at_addr

get_maximum_number_of_pgf_steward_key

Get key for maximum number of pgf stewards

get_maximum_number_of_stewards_key_at_addr

get_pgf_inflation_rate_key

Get key for inflation rate key

get_pgf_inflation_rate_key_at_addr

get_steward_inflation_rate_key

Get key for inflation rate key

get_steward_inflation_rate_key_at_addr

get_stewards_key_at_addr

is_fundings_key

Check if the given storage key is a pgf funding key.

is_fundings_key_at_addr

is_maximum_number_of_stewards_key_at_addr

is_pgf_inflation_rate_key

Check if key is a pgf inflation rate key

is_pgf_inflation_rate_key_at_addr

is_pgf_key

Check if key is inside governance address space

is_steward_inflation_rate_key

Check if key is a steward inflation rate key

is_steward_inflation_rate_key_at_addr

is_stewards_key

Check if the given storage key is a steward key. If it is, returns the steward address.

is_stewards_key_at_addr

stewards_handle

LazyMap handler for the stewards subspace

stewards_key_prefix

Obtain a storage key for stewards key