namada_governance::storage
Module keys

Governance proposal keys

Functions

get_activation_epoch_key

Get proposal activation epoch key

get_activation_epoch_key_at_addr

get_author_key

Get key of proposal author

get_author_key_at_addr

get_commit_proposal_epoch

Get the committing epoch from a proposal committing key

get_commit_proposal_id

Get the proposal id from a proposal committing key

get_commiting_proposals_prefix

Get the proposal committing key prefix

get_committing_epoch_key_at_addr

get_committing_proposals_key

Get the committing proposal key

get_content_key

Get key of proposal content

get_content_key_at_addr

get_counter_key

Get key of proposal ids counter

get_counter_key_at_addr

get_end_epoch_key_at_addr

get_funds_key

Get key of proposal funds

get_funds_key_at_addr

get_max_code_size_key_at_addr

get_max_content_key_at_addr

get_max_latency_key_at_addr

get_max_period_key_at_addr

get_max_proposal_code_size_key

Get maximum proposal code size key

get_max_proposal_content_key

Get maximum proposal content key

get_max_proposal_latency_key

Get maximum proposal latency key

get_max_proposal_period_key

Get maximum proposal period key

get_min_fund_key_at_addr

get_min_grace_epochs_key_at_addr

get_min_period_key_at_addr

get_min_proposal_fund_key

Get key for the minimum proposal fund

get_min_proposal_grace_epochs_key

Get min grace epochs proposal key

get_min_proposal_voting_period_key

Get minimum proposal period key

get_pending_key_at_addr

get_proposal_code_key

Get proposal code key

get_proposal_code_key_at_addr

get_proposal_execution_key

Get the proposal execution key

get_proposal_id

Get proposal id from key

get_proposal_key_at_addr

get_proposal_result_key

Get the proposal result key

get_proposal_type_key

Get key of a proposal type

get_proposal_type_key_at_addr

get_proposal_vote_prefix_key

Get proposal vote prefix key

get_result_key_at_addr

get_start_epoch_key_at_addr

get_vote_delegation_address

Get the delegation address from vote key

get_vote_key_at_addr

get_vote_proposal_key

Get the vote key for a specific proposal id

get_voter_address

Get voter address from vote key

get_voting_end_epoch_key

Get key of proposal voting end epoch

get_voting_start_epoch_key

Get key of proposal voting start epoch

is_activation_epoch_key

Check if key is activation epoch key

is_activation_epoch_key_at_addr

is_author_key

Check if key is author key

is_author_key_at_addr

is_balance_key

Check if key is balance key

is_commit_proposal_key

Check if key is a commit proposal key

is_committing_epoch_key_at_addr

is_content_key

Check if key is content key

is_content_key_at_addr

is_counter_key

Check if key is counter key

is_counter_key_at_addr

is_end_epoch_key

Check if key is epoch key

is_end_epoch_key_at_addr

is_funds_key_at_addr

is_governance_key

Check if key is inside governance address space

is_max_code_size_key_at_addr

is_max_content_key_at_addr

is_max_content_size_key

Check if key is a proposal max content parameter key

is_max_latency_key_at_addr

is_max_period_key_at_addr

is_max_proposal_code_size_key

Check if key is a max proposal size key

is_max_proposal_period_key

Check if key is a max proposal period param key

is_min_fund_key_at_addr

is_min_grace_epochs_key

Check if key is a min grace epochs key

is_min_grace_epochs_key_at_addr

is_min_period_key_at_addr

is_min_proposal_fund_key

Check if key is a proposal fund parameter key

is_min_proposal_voting_period_key

Check if key is a min proposal period param key

is_parameter_key

Check if key is parameter key

is_pending_key_at_addr

is_proposal_code_key

Check if key is proposal code key

is_proposal_code_key_at_addr

is_proposal_key_at_addr

is_proposal_type_key

Check if key is proposal type key

is_proposal_type_key_at_addr

is_result_key_at_addr

is_start_epoch_key

Check if key is start epoch key

is_start_epoch_key_at_addr

is_start_or_end_epoch_key

Check if key is start epoch or end epoch key

is_vote_key

Check if a key is a vote key

is_vote_key_at_addr

proposal_prefix

Get governance prefix key