namada_governance
Module utils

Governance utility functions/structs

Structs

ProposalResult

The result with votes of a proposal

ProposalVotes

Proposal structure holding votes information necessary to compute the outcome

Vote

Structure rappresenting a proposal vote

Enums

ProposalStatus

Proposal status

TallyResult

The result of a proposal

TallyType

Represents a tally type that describes the voting requirements for a proposal to pass.

Functions

compute_proposal_result

Compute the result of a proposal

is_valid_validator_voting_period

Calculate the valid voting window for a validator given proposal epoch details.

last_validator_voting_epoch

Returns the latest epoch in which a validator can vote, given the voting start and end epochs. If the pair of start and end epoch is invalid, then return None.

Type Aliases

VotePower

Alias to comulate voting power